Convert remaining utility functions from JS to TS
