+++ utils/getParentChallenge.ts
@@ -1,3 +1,5 @@
+import { ChallengeDTO } from "challenges/types/challenges.types";
+
 /**
  * 
  * @param {Number} parentId - id of the parent challenge
@@ -5,12 +7,17 @@
  * @param {Boolean} recursive - whether get the parents parent (recursive) or only the next parent; default is recursive
  * @returns 
  */
-export default function getParent(parentId, challenges, recursive = true) {
+export default function getParent(parentId: number, challenges: {[key: number]: ChallengeDTO}, recursive: boolean = true): any {
+
+   if (!recursive) {
+      // TODO
+   }
 
    try {
       let parent = challenges[parentId];
-      if (parent.parent !== parent.id && parent.id > 10) {
-         let result = getParent(parent.parent, challenges, recursive);
+      let parentParent = challenges[parent.tags.parent ?? -1];
+      if (parentParent.id !== parent.id && parseInt(parent.id) > 10) {
+         let result = getParent(parseInt(parentParent.id), challenges, recursive);
          if (result.id === 0) {
             return {
                name: "LEGACY"
@@ -20,7 +27,7 @@ export default function getParent(parentId, challenges, recursive = true) {
             name: result.name
          };
       }
-      if (parent.id === 0) {
+      if (parseInt(parent.id) === 0) {
          return {
             name: "LEGACY"
          };
