+++ utils/localStorageFunctions.ts
@@ -2,7 +2,6 @@
  * Keys used for localStorage.
  */
 export const storageKeys = {
-   titleSearch: "title-search",
    challengeFilters: "challenge-filters",
    challengeDataFilters: "challenge-data-filters",
    challengeSearch: "challenge-search",
@@ -13,38 +12,39 @@ export const storageKeys = {
 };
 
 /**
- * Get the JSON parsed value by key from sessionStorage. If no value is found,
+ * Get the JSON parsed value by key from localStorage. If no value is found,
  * a default value can optionally be used as the return.
- * @param {string} key sessionStorage key
+ * @param {string} key localStorage key
  * @param {any} defaultValue default value to return if no value was found
  * @returns {any}
  */
-export function getStorage(key, defaultValue) {
-   let value = JSON.parse(sessionStorage.getItem(key));
+export function getStorage(key: string, defaultValue: any): any {
+   let value = localStorage.getItem(key);
+   if (typeof value === "string") value = JSON.parse(value as string);
    if (value == null) value = defaultValue ?? undefined;
    return value;
 }
 
 /**
- * Sets the JSON stringified value of key into sessionStorage.
+ * Sets the JSON stringified value of key into localStorage.
  * @param {string} key
  * @param {any} value
  */
-export function setStorage(key, value) {
-   sessionStorage.setItem(key, JSON.stringify(value));
+export function setStorage(key: string, value: any) {
+   localStorage.setItem(key, JSON.stringify(value));
 }
 
 /**
- * Removes the JSON value associated with key from sessionStorage.
+ * Removes the JSON value associated with key from localStorage.
  * @param {string} key
  */
-export function removeStorage(key) {
-   sessionStorage.removeItem(key);
+export function removeStorage(key: string) {
+   localStorage.removeItem(key);
 }
 
 /**
- * Clears all values from sessionStorage. Be careful using this. ❤️
+ * Clears all values from localStorage. Be careful using this. ❤️
  */
 export function clearStorage() {
-   sessionStorage.clear();
+   localStorage.clear();
 }
