--- utils/toArray.js
@@ -1,3 +1,0 @@
-export function toArray(obj) {
-   return Object.values(obj);
-}
