+++ utils/toArray.ts
@@ -1,0 +1,3 @@
+export function toArray(obj: any): any {
+   return Object.values(obj);
+}
